Rename `attack_boards` to `lookups`
Hasn't been "just attack boards" for a very long time, if it ever even
was.
