+++ chess/src/movegen/mod.rs
@@ -1,4 +1,4 @@
-pub mod attack_boards;
+pub mod lookups;
 pub mod castling;
 pub mod legal_moves;
 pub mod moves;
